64-bit fixes for Phobos2.  All of std now passes semantic analysis on Linux with -m64.

+++ std/intrinsic.d
@@ -62,7 +62,7 @@ pure nothrow int bsf(size_t v);
  *  bsf(x21) = 0<br>
  *  bsr(x21) = 5
  */
-pure nothrow int bsr(uint v);
+pure nothrow int bsr(size_t v);
 
 /**
  * Tests the bit.
